Refactor complete and email sending works
